requirements-dev.txt
ares==0.7.1
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
chardet==4.0.0
cwe==1.6
Django==3.2.16
django-formtools==2.4
django-otp==1.1.3
django-phonenumber-field==6.1.0
django-two-factor-auth==1.14.0
djlint==1.12.1
idna==3.4
lxml==4.9.1
phonenumbers==8.12.57
psycopg2-binary==2.9.5
pydantic==1.8.2
python-dotenv>=0.19.0
pytz==2022.6
qrcode==7.3.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toolz==0.12.0
typing-extensions==3.10.0.0
urllib3==1.26.12
django-markdownify==0.9.1
xhtml2pdf==0.2.8
django-password-validators==1.7.0
django-csp==3.7
djangorestframework==3.14.0
django-tagulous==1.3.3

# temp fix to pass build, remove later when https://github.com/xhtml2pdf/xhtml2pdf/issues/589 is solved
reportlab==3.6.12

# Robot Framework dependencies
wheel
PyNaCl
Image
numpy
pyotp
robotframework
robotframework-browser
robotframework-debuglibrary
robotframework-otp
robotframework-postgresqldb
pytest
pytest-django
pytest-mock

# Include formatters and linters
-r requirements-check.txt
